refactor(NewCollectivePage): move collective query to the queries file
